refactor(core): split useChatStore into message and streaming stores (#124)
Extract useMessageStore (messages, history, autonomous insertion) and
useStreamingStore (buffers, flush timers, typing, loading) from the
monolithic useChatStore. Maintain backward-compatible facade.

Fixes #124
